Clustering functions

Functions

blm.group_firms — Method

group_firms(y,fids; nq=20,ng=10,nstart=100)

groups firms based on empirical cdf pass a vector of firm ids and a vector of wages, at the individual level nq is number of points for the empirical cdf ng is the number of groups nstart is the number of kmeans to run maxiter for each kmeans is set at 1000

Contents

Clustering functions

Functions

Contents

Index

Constraints functions

Functions

Contents

Index

Model functions

Functions

Contents

Index

blm.jl

Index

Functions

Index

blm.DistributionModel

blm.AddConstraint

blm.ConstraintAKMMono

blm.ConstraintBiggerThan

blm.ConstraintFixB

blm.ConstraintLinear

blm.ConstraintMonoK

blm.ConstraintNone

blm.ConstraintPara

blm.ConstraintSmallerThan

blm.ConstraintStationary

blm.ConstraintSum

blm.QPConstrainedSolve

blm.distributional_em!

blm.group_firms

blm.simulate