Model functions

Functions

blm.DistributionModel — Type

DistributionModel

Event study BLM model for movers - static version - dynamic version

blm.distributional_em! — Function

distributional_em!(dm::DistributionModel, data::DataFrame, maxiter::Int, tol::Float64 = 1e-8; update_level = true, constrained=nothing, iterprint=50, msg="")

Run the em algorithm on the passed DistributionalModel using the given data

blm.simulate — Method

simulate(model::DistributionModel, nn)

Using the model, simulates a dataset. The timing is important, the event happens at the end of the period, hence if a current period is u2e there will be no wage in the current period, as well as no firm.

Contents

Clustering functions

Functions

Contents

Index

Constraints functions

Functions

Contents

Index

Model functions

Functions

Contents

Index

blm.jl

Index

Functions

Index

blm.DistributionModel

blm.AddConstraint

blm.ConstraintAKMMono

blm.ConstraintBiggerThan

blm.ConstraintFixB

blm.ConstraintLinear

blm.ConstraintMonoK

blm.ConstraintNone

blm.ConstraintPara

blm.ConstraintSmallerThan

blm.ConstraintStationary

blm.ConstraintSum

blm.QPConstrainedSolve

blm.distributional_em!

blm.group_firms

blm.simulate